+++ src/lib/mympd_state.c
@@ -323,8 +323,8 @@ void partition_state_free(struct t_partition_state *partition_state) {
     FREE_SDS(partition_state->song_uri);
     FREE_SDS(partition_state->last_song_uri);
     //jukebox
-    jukebox_clear(&partition_state->jukebox_queue);
-    jukebox_clear(&partition_state->jukebox_queue_tmp);
+    jukebox_clear(&partition_state->jukebox_queue, partition_state->name);
+    jukebox_clear(&partition_state->jukebox_queue_tmp, partition_state->name);
     FREE_SDS(partition_state->jukebox_playlist);
     //lists
     list_clear(&partition_state->last_played);