+++ src/mpd_client/jukebox.h
@@ -13,8 +13,8 @@
 enum jukebox_modes jukebox_mode_parse(const char *str);
 const char *jukebox_mode_lookup(enum jukebox_modes mode);
 void jukebox_clear_all(struct t_mympd_state *mympd_state);
-void jukebox_clear(struct t_list *list);
-bool jukebox_rm_entry(struct t_list *list, long pos);
+void jukebox_clear(struct t_list *list, sds partition_name);
+bool jukebox_rm_entry(struct t_list *list, long pos, sds partition_name);
 sds jukebox_list(struct t_partition_state *partition_state, sds buffer, enum mympd_cmd_ids cmd_id,
         long request_id, long offset, long limit, sds searchstr,
         const struct t_tags *tagcols);